+++ python/primihub/FL/crypto/paillier.py
@@ -1,0 +1,23 @@
+class Paillier:
+
+    def __init__(self, public_key, private_key):
+        self.public_key = public_key
+        self.private_key = private_key
+
+    def decrypt_scalar(self, cipher_scalar):
+        return self.private_key.decrypt(cipher_scalar)
+
+    def decrypt_vector(self, cipher_vector):
+        return [self.private_key.decrypt(i) for i in cipher_vector]
+
+    def decrypt_matrix(self, cipher_matrix):
+        return [[self.private_key.decrypt(i) for i in cv] for cv in cipher_matrix]
+    
+    def encrypt_scalar(self, plain_scalar):
+        return self.public_key.encrypt(plain_scalar)
+
+    def encrypt_vector(self, plain_vector):
+        return [self.public_key.encrypt(i) for i in plain_vector]
+
+    def encrypt_matrix(self, plain_matrix):
+        return [[self.private_key.encrypt(i) for i in pv] for pv in plain_matrix]
