+++ python/primihub/FL/logistic_regression/base.py
@@ -122,7 +122,7 @@ class LogisticRegression_DPSGD(LogisticRegression):
 
         return x + noise
     
-    def compute_grad(self, x, y): 
+    def compute_grad(self, x, y):
         # compute & clip per-example gradient
         if self.multiclass:
             error = self.predict_prob(x)
@@ -149,32 +149,7 @@ class LogisticRegression_DPSGD(LogisticRegression):
         # add gaussian noise
         dw = self.add_noise(dw) / x.shape[0] + self.alpha * self.weight
         db = self.add_noise(db) / x.shape[0]
-        return dw, db 
-
-
-class Paillier:
-
-    def __init__(self, public_key, private_key):
-        self.public_key = public_key
-        self.private_key = private_key
-
-    def decrypt_scalar(self, cipher_scalar):
-        return self.private_key.decrypt(cipher_scalar)
-
-    def decrypt_vector(self, cipher_vector):
-        return [self.private_key.decrypt(i) for i in cipher_vector]
-
-    def decrypt_matrix(self, cipher_matrix):
-        return [[self.private_key.decrypt(i) for i in cv] for cv in cipher_matrix]
-    
-    def encrypt_scalar(self, plain_scalar):
-        return self.public_key.encrypt(plain_scalar)
-
-    def encrypt_vector(self, plain_vector):
-        return [self.public_key.encrypt(i) for i in plain_vector]
-
-    def encrypt_matrix(self, plain_matrix):
-        return [[self.private_key.encrypt(i) for i in pv] for pv in plain_matrix]
+        return dw, db
 
 
 class LogisticRegression_Paillier(LogisticRegression):
