+++ python/primihub/FL/logistic_regression/vfl_guest.py
@@ -3,13 +3,13 @@ from primihub.FL.utils.base import BaseModel
 from primihub.FL.utils.file import check_directory_exist
 from primihub.FL.utils.dataset import read_data, DataLoader
 from primihub.utils.logger_util import logger
+from primihub.FL.crypto.ckks import CKKS
 
 import pickle
 from sklearn.preprocessing import StandardScaler
 
 from .vfl_base import LogisticRegression_Guest_Plaintext,\
                       LogisticRegression_Guest_CKKS
-from .vfl_coordinator import CKKS
 
 
 class LogisticRegressionGuest(BaseModel):
